SDOC-41:Parts - reimplemented logic like in Paul's comment.
